requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.0
python-engineio==4.8.0
python-socketio==5.11.0
gevent==23.9.1 
gevent-websocket==0.10.1 